feat:58 디자인 수정
